test(adoption): align legacy dotnet proof contracts
